ai-assignment-brainstormer/requirements.txt
streamlit>=1.28.0
openai>=1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
Pillow>=10.0.0
pytesseract>=0.3.10
